Create LBH-AG v1.2.3.lsl
The anti-grief shown works against both large AND recursive damage sources without filling up the list with each individual hit. This is, obviously, slower and will more likely drops hits during high latency BUT it should be very rare if not virtually impossible for it to stack-heap under stressful conditions. However since damage is still updated at the time of impact and instead of waiting on the timer event, it should overall perform better in a majority of situations.