v1.3 'Fixed-Location' Update
The intent is to grab the source's location with llDetectedPos during the collision event and using that as a reference. If no such reference exists, the system will just default to the current method. 

The goal is to clamp how the list is generated so it doesn't cause the system to stack-heap when exposed to many damage sources at a time. The eldest entries in the list will be dropped once it fills.