Update LBA-v.2.3.lsl
Discovered a bug, if owner and src are the same listfindlist doesn't work. forgot an else @ line 144, if attached do this, else go through source finding.
it was trying to find the source of an attachment which is the owner, so recent was [owner,owner,restoftheshit], and llListFindList(test,[same,same]); doesn't work for some weird reason.